Enable user to keep stream after it finishes
This is necessary if the stream servers multiple purposes, such as also
being used as a `Sink`. The user may wish to finish flushing response
data on a TCP connection even if the other end has performed a shutdown
of their write end.

This change means that streams are no longer automatically dropped when
they finish (by necessity), so a new `#[must_use]` type is introduced
that makes it clear that the user needs to explicitly drop the stream.
Unfortunately, it does not produce a warning in the most common usage
scenario (when used in a pattern match) due to
https://github.com/rust-lang/rust/issues/66425.

This patch also adds the ability to remove a stream from
`StreamUnordered`. If the stream type is `Unpin`, the stream can even be
returned.