Problem: StreamUnordered::push naming
By analogy with `Vec::push` it suggests that streams are append to the
end of the list. However, if you use `iter_mut()` you'll see that
streams are iterated over in reverse order.

Upon reading internal comments, it became clear that `StreamUnordered`
is backed with linked lists and inserts new streams at the head.

Solution: rename `push` to `push_front` and deprecate `push`
(naming is by analogy with `VecDeque`)

Bumps up the version to 0.6.0